IOS-6028 Extract DiscussionNotificationMode to its own file
